### 2017-03-18 #213 @whedon
[PRE REVIEW]: OSMnx: A Python package to work with graph-theoretic OpenStreetMap street networks
**Submitting author:** gboeing (<a href="http://orcid.org/0000-0003-1851-6411">Geoff Boeing</a>)
**Repository:** <a href="https://github.com/gboeing/osmnx/" target ="_blank">https://github.com/gboeing/osmnx/</a>
**Version:** v0.4

Thanks for submitting your paper to JOSS gboeing. The JOSS editor (shown at the top of this issue) will work with you on this issue to find a reviewer for your submission before creating the main review issue.
