### 2017-03-19 #213 @arokem
@rdottle : would you like to review this software paper for JOSS?
