### 2017-03-19 #213 @arokem
Awesome! Thank you.

Take a look at the guidelines for reviewers here: http://joss.theoj.org/about#reviewer_guidelines

I will start the review on another issue, where there will be a check-list for you to go through, to make sure that the software meets all the requirements for publication. If there are items on the checklist that are not currently met by the software, post an issue on the software project's Github repo. When everything is place, give me the 👍  on the review issue.

As for timeline -- if you can review this within 2 weeks or so, that would be great. Let me know if you think it will take much longer than that.
