### 2017-03-19 #215 @whedon
[REVIEW]: OSMnx: A Python package to work with graph-theoretic OpenStreetMap street networks
**Submitting author:** @gboeing (<a href="http://orcid.org/0000-0003-1851-6411">Geoff Boeing</a>)
**Repository:** <a href="https://github.com/gboeing/osmnx/" target ="_blank">https://github.com/gboeing/osmnx/</a>
**Version:** v0.4

- [x] **Authorship:** Has the submitting author (gboeing) made major contributions to the software?
