+++ osmnx/__init__.py
@@ -14,4 +14,4 @@ from .stats import *
 from .utils import *
 from .buildings import *
 
-__version__ = '0.4.1dev'
+__version__ = '0.4.1'
