### 2017-04-07 3e8a1fa
add get node colors by attribute value
