### 2017-04-07 680429b
add elevation module to get node elevations from google api
