### 2017-04-07 a705806
remove elevation tests
