### 2017-04-07 23739cb
calculate edge grades and their absolute values, and add get_route_edge_attributes function
