### 2017-04-07 92bab0a
increase coverage and skip elevation module as it requires api keys
