### 2017-04-07 cd3684e
increase coverage and skip elevation module as it requires api keys
