+++ tests/test_osmnx.py
@@ -94,6 +94,9 @@ def test_plots():
     
     G = ox.graph_from_place('Piedmont, California, USA', network_type='drive', simplify=False)
     G2 = ox.simplify_graph(G, strict=False)
+    nc = ox.get_node_colors_by_attr(G2, 'osmid')
+    ec = ox.get_edge_colors_by_attr(G2, 'length')
+    
     fig, ax = ox.plot_graph(G, save=True, file_format='png')
     
     G_simplified = ox.simplify_graph(G)
@@ -121,6 +124,9 @@ def test_routing_folium():
     origin_node = ox.get_nearest_node(G, origin)
     destination_node = ox.get_nearest_node(G, destination)
     route = nx.shortest_path(G, origin_node, destination_node)
+    
+    attributes = ox.get_route_edge_attributes(G, route, 'length')
+    
     fig, ax = ox.plot_graph_route(G, route, save=True, filename='route', file_format='png')
     fig, ax = ox.plot_graph_route(G, route, origin_point=origin, destination_point=destination,
                                   save=True, filename='route', file_format='png')
@@ -134,15 +140,4 @@ def test_buildings():
     gdf = ox.buildings_from_place(place='Piedmont, California, USA')
     gdf = ox.buildings_from_address(address='San Francisco, California, USA', distance=300)
     fig, ax = ox.plot_buildings(gdf)
-    
 
-#def test_elevation():
-    #
-    #from keys import google_elevation_api_key
-    #G = ox.graph_from_place('Piedmont, California, USA', network_type='drive')
-    #G = ox.add_node_elevations(G, api_key=google_elevation_api_key)
-    #G = ox.add_edge_grades(G)
-    #nc = ox.get_node_colors_by_attr(G, 'elevation')
-    #ec = ox.get_edge_colors_by_attr(G, 'grade')
-    
-    
