+++ docs/source/conf.py
@@ -31,15 +31,14 @@ autodoc_mock_imports = [
     'requests',
     'geopandas',
     'networkx',
+    'fiona',
     'matplotlib.pyplot',
     'matplotlib.cm',
     'matplotlib.collections',
     'shapely',
     'shapely.ops',
     'shapely.geometry',
-    'descartes',
-    'geopy.distance',
-    'geopy.geocoders'
+    'descartes'
 ]
 
 # -- General configuration ------------------------------------------------
@@ -76,9 +75,9 @@ author = 'Geoff Boeing'
 # built documents.
 #
 # The short X.Y version.
-version = '0.3'
+#version = '0.3'
 # The full version, including alpha/beta/rc tags.
-release = '0.3'
+#release = '0.3'
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
