+++ docs/source/index.rst
@@ -1,7 +1,4 @@
-.. OSMnx documentation master file, created by
-   sphinx-quickstart on Tue Jan 31 15:11:18 2017.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
+
 
 OSMnx documentation
 ===================
