### 2017-04-12 08f0f68
add color sequence functions
