### 2017-04-14 ea42b69
handle errors if polygon has no area or is invalid
