+++ osmnx/core.py
@@ -35,21 +35,21 @@ from .stats import count_streets_per_node
 
 def save_to_cache(url, response_json):
     """
-    Save an HTTP response json object to the cache. 
-    
-    If the request was sent to server via POST instead of GET, 
+    Save an HTTP response json object to the cache.
+
+    If the request was sent to server via POST instead of GET,
     then URL should be a GET-style representation of request. Users should always pass OrderedDicts instead of dicts
     of parameters into request functions, so that the parameters stay in the same order each time, producing the same URL string,
     and thus the same hash. Otherwise the cache will eventually contain multiple saved responses for the same
     request because the URL's parameters appeared in a different order each time.
-    
+
     Parameters
     ----------
     url : string
         the url of the request
     response_json : dict
         the json response
-    
+
     Returns
     -------
     None
@@ -57,32 +57,32 @@ def save_to_cache(url, response_json):
     if globals.use_cache:
         if response_json is None:
             log('Saved nothing to cache because response_json is None')
-        else:        
+        else:
             # create the folder on the disk if it doesn't already exist
             if not os.path.exists(globals.cache_folder):
                 os.makedirs(globals.cache_folder)
 
-            # hash the url (to make filename shorter than the often extremely long url) 
+            # hash the url (to make filename shorter than the often extremely long url)
             filename = hashlib.md5(url.encode('utf-8')).hexdigest()
             cache_path_filename = '{}/{}.json'.format(globals.cache_folder, filename)
-            
+
             # dump to json, and save to file
             json_str = make_str(json.dumps(response_json))
             with io.open(cache_path_filename, 'w', encoding='utf-8') as cache_file:
                 cache_file.write(json_str)
-            
+
             log('Saved response to cache file "{}"'.format(cache_path_filename))
-        
+
 
 def get_from_cache(url):
     """
     Retrieve a HTTP response json object from the cache.
-    
+
     Parameters
     ----------
     url : string
         the url of the request
-    
+
     Returns
     -------
     response_json : dict
@@ -96,20 +96,20 @@ def get_from_cache(url):
         if os.path.isfile(cache_path_filename):
             response_json = json.load(io.open(cache_path_filename, encoding='utf-8'))
             log('Retrieved response from cache file "{}" for URL "{}"'.format(cache_path_filename, url))
-            return response_json        
+            return response_json
 
 
 def get_pause_duration(recursive_delay=5, default_duration=10):
     """
     Check the Overpass API status endpoint to determine how long to wait until next slot is available.
-    
+
     Parameters
     ----------
     recursive_delay : int
         how long to wait between recursive calls if server is currently running a query
     default_duration : int
         if fatal error, function falls back on returning this value
-    
+
     Returns
     -------
     int
@@ -144,14 +144,14 @@ def get_pause_duration(recursive_delay=5, default_duration=10):
             # any other status is unrecognized - log an error and return default duration
             log('Unrecognized server status: "{}"'.format(status), level=lg.ERROR)
             return default_duration
-            
+
     return pause_duration
 
 
 def nominatim_request(params, pause_duration=1, timeout=30, error_pause_duration=180):
     """
     Send a request to the Nominatim API via HTTP GET and return the JSON response.
-    
+
     Parameters
     ----------
     params : dict or OrderedDict
@@ -162,21 +162,21 @@ def nominatim_request(params, pause_duration=1, timeout=30, error_pause_duration
[…]
