+++ osmnx/save_load.py
@@ -23,7 +23,7 @@ from .utils import log, make_str
 def save_gdf_shapefile(gdf, filename=None, folder=None):
     """
     Save GeoDataFrame as an ESRI shapefile.
-    
+
     Parameters
     ----------
     gdf : GeoDataFrame
@@ -32,33 +32,33 @@ def save_gdf_shapefile(gdf, filename=None, folder=None):
         what to call the shapefile (file extensions are added automatically)
     folder : string
         where to save the shapefile, if none, then default folder
-    
+
     Returns
     -------
     None
     """
     if folder is None:
         folder = globals.data_folder
-    
+
     if filename is None:
         filename = make_shp_filename(gdf.gdf_name)
-    
+
     # give the save folder a filename subfolder to make the full path to the files
     folder_path = '{}/{}'.format(folder, filename)
-    
+
     # if the save folder does not already exist, create it with a filename subfolder
     if not os.path.exists(folder_path):
         os.makedirs(folder_path)
     gdf.to_file(folder_path)
     if not hasattr(gdf, 'gdf_name'):
         gdf.gdf_name = 'unnamed'
-    log('Saved the GeoDataFrame "{}" as shapefile "{}"'.format(gdf.gdf_name, folder_path)) 
-    
-    
+    log('Saved the GeoDataFrame "{}" as shapefile "{}"'.format(gdf.gdf_name, folder_path))
+
+
 def save_graph_shapefile(G, filename='graph', folder=None, encoding='utf-8'):
     """
     Save graph nodes and edges as ESRI shapefiles to disk.
-    
+
     Parameters
     ----------
     G : networkx multidigraph
@@ -68,7 +68,7 @@ def save_graph_shapefile(G, filename='graph', folder=None, encoding='utf-8'):
         the folder to contain the shapefiles, if None, use default data folder
     encoding : string
         the character encoding for the saved shapefiles
-    
+
     Returns
     -------
     None
@@ -77,17 +77,17 @@ def save_graph_shapefile(G, filename='graph', folder=None, encoding='utf-8'):
     start_time = time.time()
     if folder is None:
         folder = globals.data_folder
-    
+
     # convert directed graph G to an undirected graph for saving as a shapefile
     G_save = G.copy()
     G_save = get_undirected(G_save)
 
-    
+
     # create a GeoDataFrame of the nodes and set CRS
     nodes = {node:data for node, data in G_save.nodes(data=True)}
     gdf_nodes = gpd.GeoDataFrame(nodes).T
     gdf_nodes.crs = G_save.graph['crs']
-    
+
     # create a geometry column then drop the x and y columns
     gdf_nodes['geometry'] = gdf_nodes.apply(lambda row: Point(row['x'], row['y']), axis=1)
     gdf_nodes = gdf_nodes.drop(['x', 'y'], axis=1)
@@ -96,7 +96,7 @@ def save_graph_shapefile(G, filename='graph', folder=None, encoding='utf-8'):
     gdf_nodes['osmid'] = gdf_nodes['osmid'].astype(np.int64)
     for col in [c for c in gdf_nodes.columns if not c == 'geometry']:
         gdf_nodes[col] = gdf_nodes[col].fillna('').map(make_str)
-    
+
     # create a list to hold our edges, then loop through each edge in the graph
     edges = []
     for u, v, key, data in G_save.edges(keys=True, data=True):
@@ -111,7 +111,7 @@ def save_graph_shapefile(G, filename='graph', folder=None, encoding='utf-8'):
             point_u = Point((G_save.node[u]['x'], G_save.node[u]['y']))
             point_v = Point((G_save.node[v]['x'], G_save.node[v]['y']))
[…]
