+++ osmnx/globals.py
@@ -27,4 +27,3 @@ log_filename = 'osmnx'
 # useful osm tags - note that load_graphml expects a consistent set of tag names for parsing
 useful_tags_node = ['ref', 'highway']
 useful_tags_path = ['bridge', 'tunnel', 'oneway', 'lanes', 'ref', 'name', 'highway', 'maxspeed', 'service', 'access', 'area', 'landuse', 'width', 'est_width']
-

[…]
