### 2017-04-17 ba1d1b3
fixes #46
