### 2017-04-19 #215 @rdottle
@gboeing
After review, the only issue found was missing contributing guidelines, which I opened as an issue in the repo.
