### 2017-04-19 6773221
fixes #47
