+++ CONTRIBUTING.md
@@ -1,0 +1,17 @@
+Thanks for using OSMnx and for considering contributing to it!
+
+#### If you've found an error:
+
+  - read the error message and [documentation](https://osmnx.readthedocs.io/)
+  - search through the [open issues](https://github.com/gboeing/osmnx/issues?q=is%3Aopen+is%3Aissue) and [closed issues](https://github.com/gboeing/osmnx/issues?q=is%3Aissue+is%3Aclosed) first
+  - if the problem is with a dependency of this project, open an issue on the dependency's repo
+  - if the problem is with OSMnx and you can fix it simply, please submit a PR
+  - if the problem persists, please open an issue in the [issue tracker](https://github.com/gboeing/osmnx/issues) including a minimal working example to reproduce the problem
+
+#### If you have a feature proposal or want to contribute
+
+  - post your proposal on the [issue tracker](https://github.com/gboeing/osmnx/issues) so we can review it together (some proposals may not be a good fit for the project)
+  - fork the repo, make your change (adhering to OSMnx's existing coding, commenting, and docstring styles), [test it](https://github.com/gboeing/osmnx/tree/master/tests), and submit a PR
+  - respond to code review
+ 
+Every piece of software is a work in progress. This project is the result of many hours of work contributed freely by myself and the many people that build the projects it depends on. Thank you for contributing!
