### 2017-04-19 #215 @gboeing
Thanks @rdottle, contributing guidelines have just been added to the repo.
