### 2017-04-22 6e10853
fix get_undirected to handle uncommon cases of not retaining parallel edges
