### 2017-04-22 155b142
update graph_from_place docstring
