### 2017-04-22 673ddf2
fix graph_from_place docstring
