### 2017-04-23 8ab104f
update plot docstrings
