### 2017-04-24 #215 @arokem
Thanks @rdottle!

@gboeing : would it be reasonable to make a release of the current state of the software, and upload an archive of that to Zenodo? We still need a DOI/archive for the software to publish.
