### 2017-04-24 #215 @arfon
@gboeing - it looks like the paper compilation is failing as you're not actually citing the papers listed in `paper.bib` in the `paper.md` text (see this page for example http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html).

Also, could you update the paper metadata with the affiliations formatted correctly?:

```
authors:
  - name: Geoff Boeing
    orcid: 0000-0003-1851-6411
    affiliation: 1
affiliations:
  - name: University of California, Berkeley
    index: 1
```
