### 2017-04-24 9bfddd0
update citations
