### 2017-04-24 #215 @gboeing
@arfon paper citations and affiliations are updated
