### 2017-04-24 #215 @arfon
@rdottle many thanks for your review here and to @arokem for editing this one ✨

@gboeing - your paper is now accepted into JOSS and your paper DOI is http://dx.doi.org/10.21105/joss.00215 :zap: 🚀 💥
